Refactor: 에디터 디렉토리 구조 변경
각 셀에 대한 컴포넌트 파일들을 하나로 묶어 cells 디렉토리로 생성
각 셀 컴포넌트 파일들의 이름을 디렉토리로 생성하고 index로 이름 변경
  - 이름 변경에 따른 파일 패스 전부 변경
